+++ agent/dockerclient/dockerauth/dockerauth_test.go
@@ -22,7 +22,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/registry"
 )
 
 type authTestPair struct {
@@ -122,7 +122,7 @@ func TestAuthErrors(t *testing.T) {
 	for _, pair := range badPairs {
 		provider := NewDockerAuthProvider(pair.t, []byte(pair.a))
 		result, _ := provider.GetAuthconfig("nginx", nil)
-		if !reflect.DeepEqual(result, types.AuthConfig{}) {
+		if !reflect.DeepEqual(result, registry.AuthConfig{}) {
 			t.Errorf("Expected empty auth config for %v; got %v", pair, result)
 		}
 	}
@@ -132,7 +132,7 @@ func TestAuthErrors(t *testing.T) {
 func TestEmptyConfig(t *testing.T) {
 	provider := NewDockerAuthProvider("", []byte(""))
 	authConfig, _ := provider.GetAuthconfig("nginx", nil)
-	if !reflect.DeepEqual(authConfig, types.AuthConfig{}) {
+	if !reflect.DeepEqual(authConfig, registry.AuthConfig{}) {
 		t.Errorf("Expected empty authconfig to not return any auth data at all")
 	}
 }