+++ agent/dockerclient/dockerauth/ecr_test.go
@@ -30,7 +30,7 @@ import (
 	mock_async "github.com/aws/amazon-ecs-agent/ecs-agent/async/mocks"
 	"github.com/aws/amazon-ecs-agent/ecs-agent/credentials"
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/registry"
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -122,7 +122,7 @@ func TestGetAuthConfigNoMatchAuthorizationToken(t *testing.T) {
 
 	authconfig, err := provider.GetAuthconfig(proxyEndpoint+"/myimage", registryAuthData)
 	require.Error(t, err, "Expected error if the proxy does not match")
-	assert.Equal(t, types.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
+	assert.Equal(t, registry.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
 }
 
 func TestGetAuthConfigBadBase64(t *testing.T) {
@@ -157,7 +157,7 @@ func TestGetAuthConfigBadBase64(t *testing.T) {
 
 	authconfig, err := provider.GetAuthconfig(proxyEndpoint+"/myimage", registryAuthData)
 	require.Error(t, err, "Expected error to be present, but was nil", err)
-	assert.Equal(t, types.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
+	assert.Equal(t, registry.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
 }
 
 func TestGetAuthConfigMissingResponse(t *testing.T) {
@@ -190,7 +190,7 @@ func TestGetAuthConfigMissingResponse(t *testing.T) {
 		t.Fatal("Expected error to be present, but was nil", err)
 	}
 	require.Error(t, err, "Expected error to be present, but was nil", err)
-	assert.Equal(t, types.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
+	assert.Equal(t, registry.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
 }
 
 func TestGetAuthConfigECRError(t *testing.T) {
@@ -220,7 +220,7 @@ func TestGetAuthConfigECRError(t *testing.T) {
 
 	authconfig, err := provider.GetAuthconfig(proxyEndpoint+"/myimage", registryAuthData)
 	require.Error(t, err, "Expected error to be present, but was nil", err)
-	assert.Equal(t, types.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
+	assert.Equal(t, registry.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
 }
 
 func TestGetAuthConfigNoAuthData(t *testing.T) {
@@ -235,7 +235,7 @@ func TestGetAuthConfigNoAuthData(t *testing.T) {
 
 	authconfig, err := provider.GetAuthconfig(proxyEndpoint+"/myimage", nil)
 	require.Error(t, err, "Expected error to be present, but was nil", err)
-	assert.Equal(t, types.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
+	assert.Equal(t, registry.AuthConfig{}, authconfig, "Expected Authconfig to be empty, but was %v", authconfig)
 }
 
 func TestIsTokenValid(t *testing.T) {